+++ ScriptsUI Panels/Toolbox.jsx
@@ -14,7 +14,7 @@
     #include "Toolbox_Assets/HelperScripts/TOOL_PathReformatter.jsx";
 
     var ToolboxData = new Object();
-    var version = "2.2.12";
+    var version = "2.2.13";
     var scriptFile = new File($.fileName);
     var scriptPath = scriptFile.parent.fsName;
     var systemFont = "";
@@ -32,6 +32,7 @@
     var currentProjectToGFXPath = "";
     var currentProjectAssetsPath = "";
     var currentProjectSFPath = "";
+    var importAssetsInput = "";
     var projectSelection = 0;
     var projectCode = "";
     var showHideProject = false;
@@ -1947,10 +1948,15 @@
             }
 
             pal.gr_one.cmds1.textField.onChanging = function(){
+                importAssetsInput = this.text;
                 if(pal.gr_one.cmds1.textField.text == ""){
                     pal.layout.layout(true);
                     pal.layout.resize();
                 }
+            }
+
+            pal.gr_one.cmds1.textField.onChange = function(){
+                importAssetsInput = this.text;
             }
 //
             // pal.gr_one.cmds1.pastePaths.onClick = pasteFilePaths;
@@ -1959,12 +1965,16 @@
             pal.gr_one.cmds1.textField.onActivate = function(){
                 if(pal.gr_one.cmds1.textField.text ==  "Paths to files go here..."){
                     pal.gr_one.cmds1.textField.text = "";
+                    importAssetsInput = "";
                 }
             }
 
             pal.gr_one.cmds1.textField.onDeactivate = function(){
                 if(pal.gr_one.cmds1.textField.text == ""){
                     pal.gr_one.cmds1.textField.text =  "Paths to files go here...";
+                    importAssetsInput = "";
+                } else {
+                    importAssetsInput = pal.gr_one.cmds1.textField.text;
                 }
             }
 
@@ -1994,10 +2004,7 @@
             var consdupsBtn = pal.gr_three.cmds1.consolDups;
             consdupsBtn.value = true;
 
-            pal.gr_one.cmds1.ImportPaths.onClick = function() {
-                var field = this.parent ? this.parent.textField : null;
-                importFilesFromPaths(field ? field.text : pal.gr_one.cmds1.textField.text);
-            };
+            pal.gr_one.cmds1.ImportPaths.onClick = importFilesFromPaths;
             pal.gr_one.cmds1.ImportPaths.helpTip = "Paste one absolute file path per line, or a folder path followed by filenames. File URLs and configured Mac/Windows root paths are supported.";
             pal.gr_one.cmds4.SourceFromRen.onClick = importSourceProjectsFromRenDialoge;
             pal.gr_one.cmds4.SourceFromRen.helpTip = "Select quicktime files in project window and click. This will import the AE project used to created the selected the quicktimes. The selected quicktimes file must have been rendered from After Effects with embedded metadata.";
@@ -4188,12 +4195,15 @@ function removeText(s){
 
 ////IMPORT SOURCE FILES FROM INPUT TEXT PATHS - ONCLICK BUTTON ACTION///////
 
-    function importFilesFromPaths(paths) {
+    function importFilesFromPaths() {
         try {
             getCurrentDate();
             appendLog("Function_Tracking", currentDateYMD + " " + userName + " // :: Function Name: importFilesFromPaths :: // ", "Log");
         } catch (logError) {}
         var options = parseBuildOptionsToArr();
+        var paths = "";
+        if (this && this.parent && this.parent.textField) paths = this.parent.textField.text;
+        if ((!paths || paths === "Paths to files go here...") && importAssetsInput) paths = importAssetsInput;
         return toolboxImportAssets(paths, options[6], options[7]);
     }